backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
ascii_magic
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-grpc
